Feat: change price from int to float64 on product model

+++ internal/models/product.model.go
@@ -7,7 +7,7 @@ type Product struct {
 	CategoryId      int       `json:"category_id" db:"category_id"`
 	Name            string    `json:"name" db:"name"`
 	Description     string    `json:"description" db:"description"`
-	Price           int       `json:"price" db:"price"`
+	Price           float64       `json:"price" db:"price"`
 	Stock           int       `json:"stock" db:"stock"`
 	CreatedAt       time.Time `json:"created_at" db:"created_at"`
 	UpdatedAt       time.Time `json:"updated_at" db:"updated_at"`
